+++ src/components/stat/stat.tsx
@@ -22,6 +22,7 @@ import React, {
   HTMLAttributes,
   FunctionComponent,
   ReactNode,
+  createElement,
 } from 'react';
 import { CommonProps, keysOf } from '../common';
 import classNames from 'classnames';
@@ -82,6 +83,14 @@ export interface EuiStatProps {
    * Size of the title. See EuiTitle for options ('s', 'm', 'l'... etc)
    */
   titleSize?: EuiTitleSize;
+  /**
+   * HTML Element to be used for title
+   */
+  titleElement?: string;
+  /**
+   * HTML Element to be used for description
+   */
+  descriptionElement?: string;
 }
 
 export const EuiStat: FunctionComponent<
@@ -96,6 +105,8 @@ export const EuiStat: FunctionComponent<
   title,
   titleColor = 'default',
   titleSize = 'l',
+  titleElement = 'p',
+  descriptionElement = 'p',
   ...rest
 }) => {
   const classes = classNames(
@@ -112,21 +123,32 @@ export const EuiStat: FunctionComponent<
     }
   );
 
+  const commonProps = {
+    'aria-hidden': true,
+  };
+
   const descriptionDisplay = (
     <EuiText size="s" className="euiStat__description">
-      <p aria-hidden="true">{description}</p>
+      {createElement(descriptionElement, commonProps, description)}
     </EuiText>
   );
 
+  const titlePropsWithColor = {
+    'aria-hidden': true,
+    style: {
+      color: `${titleColor}`,
+    },
+  };
+
+  const titleChildren = isLoading ? '--' : title;
+
   const titleDisplay = isColorClass(titleColor) ? (
     <EuiTitle size={titleSize} className={titleClasses}>
-      <p aria-hidden="true">{isLoading ? '--' : title}</p>
+      {createElement(titleElement, commonProps, titleChildren)}
     </EuiTitle>
   ) : (
     <EuiTitle size={titleSize} className={titleClasses}>
-      <p aria-hidden="true" style={{ color: `${titleColor}` }}>
-        {isLoading ? '--' : title}
-      </p>
+      {createElement(titleElement, titlePropsWithColor, titleChildren)}
     </EuiTitle>
   );
 
@@ -149,7 +171,9 @@ export const EuiStat: FunctionComponent<
       {!reverse && descriptionDisplay}
       {titleDisplay}
       {reverse && descriptionDisplay}
-      {screenReader}
+      {typeof title === 'string' &&
+        typeof description === 'string' &&
+        screenReader}
     </Fragment>
   );